Moving decorators into a sub-directory
As I was looking at Sipity's documentation at
http://www.rubydoc.info/github/ndlib/sipity/master, I was noticing
that all of the decorators are in the main namespace. Which means that
there are, at this point 5 concepts in the main Sip namespace that
may not be immediately obvious.

So I've opted to move the decorators into a sub-directory. This
then creates a logical place to document the Decorators module.