Moving Forms into subdirectory
As I was looking at Sipity's documentation at
http://www.rubydoc.info/github/ndlib/sipity/master, I was noticing
that all of the forms are in the main namespace. This meant that there
were quite a few concepts being exposed without proper introduction.

So I've opted to move the `forms/sip/forms` into a sub-directory. This
then creates a logical place to document the Sip::Forms module and
provide a proper introduction to the concepts.