Moving Runners into a subdirectory
As I was looking at Sipity's documentation at

http://www.rubydoc.info/github/ndlib/sipity/master, I was noticing
that all of the runners are in the main namespace. This meant that
there were quite a few concepts being exposed without proper
introduction.

So I've opted to move the runners into `runners/sip/runners` into.
This then creates a logical place to document the Sip::Runners module
and provide a proper introduction to the concepts.