Adding Sipity::Models::Header#processing_state
What is a processing_state? The SIP is working its way to becoming an
AIP. I'm envisioning a singular workflow for a given header. The
workflow is responsible for moving the data towards completion. And
we persist the processing_state of the data as its moving to and from
disk.