Renaming module Repo to RepositoryMethods
I'm interested in privatizing, as a good capitalist, all of the
underlying RepositoryMethod constants. The reason being that I want
to start crafting support module functions that are usable within the
Repository but not outside.

I believe this highlights a possible code smell, but for now I'm
willing to live with that as I'm hoping to isolate the stink to the
Repository class.

It is also possible that I'm going to need to create multiple
repositories at a future point. That would be the likely refactoring.