Factoring away deprecated methods
I'm in an ongoing pursuit to tease apart the repository. My
understanding of the repository concept is improving. This is amongst
the first steps.

Further plans are to separate commands and queries. Queries can be
public modules that decorators and so on can include.

This is my ongoing effort to avoid over-reliance on Rails. I can treat
it as a Repository layer, but I don't want to do that (yet).