Adding #to_model for Form
The form object, when processed by the form_builder was using the
polymorphic URL builder. With the following change to Rails (4.1.8 to
4.2.0), the polymorphic path could not be inferred.

I believe the following commit to Rails to be the "culprit" but I
both welcome this change and am not certain this is the true reason
of failure. However the tests now pass.

See rails/rails@b89dcde8