Moving views into a controllers subdirectory
While it may seem redundant to have an app/views/sipity/controllers
directory, I was running into problems related to the controller path.
I believe this may be a method that I should not override.

Below is part of the message I was getting:

"Missing template sipity/controllers/citations/new"