Adding work_type to Sip
At this point, the system is being built with ETDs in mind. However,
other work types (i.e. Senior Thesis) may very well enter into the
application. I want to begin laying the ground work for that.

At present, I'm envisioning an enum, as that is the quickest thing to
implement. If Sipity is to become an application for everyone, its
going to get more complicated as we proceed. But that is a problem for
tomorrow.