Renaming role to acting_as
The word "role" comes with so much baggage. In conversations, it can
easily be mixed with the idea of a group (i.e. Catalogers can be
easily perceived as a Group and a Role).

This is unlikely to be the final name, but it creates a new word that
allows for easier future changes.

In an effort to better reflect the intention of the parameter and
methods, I'm renaming a few of them as well.

NOTE: This is a major change, you will need to drop your development
and test database and load the new schema:

```console
rake db:drop db:create db:schema:load
```