Removing collaborators from Sip Creation
The initial SIP creation page will have:

* Title
* Work Type
* Publication Strategy
* Access Rights

This means that I'm no longer concerned about the collaborators on
this page; It also means that I can remove some of the nasty nested
forms behavior. But for now that is not the expected behavior.

Some of this behavior may come back, but at this point it is not
needed.