Create AccessRight model
I opted to create a separate model as both a Sip and an
associated Attachment could have different access rights. I also went
the route of implementing a date range to reflect that as we deal with
Embargo's and Lease's the data gets convoluted as we attempt to
reflect two concepts (rights and dates) in one user input.