Adding Transient Answer model
There are some answers that we ask our patrons to complete that are
prompts for additional thought or consideration. This model is the
persistence of those answers, and is to be used to continue prompting
for more permanent information.

On the initial page for creating a sip we want to ask the following:
What is the applicable Access Right?

Possible Answers:
  * Open Access
  * Restricted
  * Private
  * It will change over time

In the case of Open Access, Restricted, and Private the answer is very
simple; We make an entry in AccessRight that starts as of today for
the given answer.

However, if the answer is "It will change over time", we don't want
to present additional decision points immediately, but instead want
to persist an answer that will require additional follow-up as we
assist our patron in completing their SIP.

The follow-up to the answer would allow for us to introduce a complex
decision in a more user focused and "gentle" manner. The idea being
that the process of depositing something should be a conversation.