Extracting methods from decorator to form
These methods do not make sense as part of the presenter decoration
for a Sip's; They are relevant for form rendering and validation and
not needed for display.

I also do not believe that they will be part of the edit a sip action.
The edit behavior is going to change to reflect that we are going to
be stepping our patrons through the deposit much more than presenting
a gigantic form to input lots of data.