Rendering EntityEnrichmentAction in show view
EntityEnrichmentActions are going to vary by:

* User - some users may not need to or be able to see a certain option
  a grad may not be given the opportunity to add all kinds of
  metadata to their deposit, that may be reserved for an approving
  librarian or metadata specialist)
* State - based on an object's state, additional enrichments may not
  be available; Once an ETD has been approved for review we may choose
  to disallow one or more forms, opting instead to make the "kitchen
  sink" enrichment option available.
* Entity Type - the enrichment options for an ETD vs. an Image vs.
  some other work type will vary

An important thing as we continue developing this application is to
make sure that we have named concepts. In naming things we can grow
our understanding as a team. It also draws attention to the fact that
most everything should have a proper name.