Adjusting ETD Student Submission as per meeting
This is a prescriptive adjustment to the state machine. That is to say
it is describing the workflow as it should be. It has not been
finalized by Shari, so it should still be considered a draft.

> Jeremy will update the state diagram to include the Advisor signoff
> before the "under review" state

See https://docs.google.com/a/nd.edu/document/d/1Gvf4A71SGBy30FcBcaMnfUeefMgCMJnCzdQ1P3owN7Q/edit

[skip ci]