Preparing for next enrichment
Building on DLTP-346 and preparing for DLTP-347, I'm looking to reduce
the need for multiple controllers based on enrichment.

This current step is moving the routing to a less resource-based route
and more of a task-based route.

From:

```ruby
resources :works do
  resource :work_description
end
```

To:

```ruby
get(
  'works/:work_id/describe',
  to: 'work_descriptions#new', as: 'describe_work'
)
post 'works/:work_id/describe', to: 'work_descriptions#create'
```

My eventual goal for describe and attach are to leverage a shared
route as follows:

```ruby
get(
  'works/:work_id/:enrichment_type',
  to: 'work_descriptions#new', as: 'enrich_work'
)
post 'works/:work_id/:enrichment_type', to: 'work_enrichments#update'
```

Note that the action names change from #new/#update to #edit/#create.
This is because an enrichment is always a modification to an
underlying work.

For example (and it is arbitrary):

One enrichment is for DOIs:

* Title
* Publisher
* Publication Year
* Authors

Another enrichment is for Publication Intent:

* Publication date
* Publisher
* Publication format

If you complete the Publication Intent enrichment, then go to the DOI
enrichment, the publisher and publication year would be pre-populated
because you have already provided some of the information.