Handling enrichment type for work enrichments
Given that there are several varieties of enrichments, I want to
expose a common pluggable location for those enrichments. This
ultimately means that there can be different enrichments for a given
work (i.e. Advisory Councel is relevant to an ETD but not a Document).
It also means that for a given enrichment name, the attributes could
vary by work type (i.e. Describe for ETD would include Abstract, for
an Image it wouldn't have an Abstract but instead a Description).