Marking up our action via Schema.org
As part of an experiment we are attempting to reduce the collision of
concerns related to CSS, Javascript, and automated testing.

Dan will be working to articulate the markup guidelines; At present
we are looking to reserve dom classes for CSS and JS interaction.
Testing is a somewhat different concern. The hope is that by spending
time marking up via Schema.org, when someone sees that markup they
will consider "What does this mean?" and be less prone to
modification.

REVIEW: This is an experimental idea; We are using the Schema.org
structure to drive our automated testing. This should be viewed as
experimental and will be part of an ongoing conversation amongst the
developers.