Creating a sipity:rebuild_interfaces task
This is an experimental proposal, but one that I want to start kicking
around.

The intent is to create an object that has the same method signature
of existing repository methods; This way we can leverage a mock
repository that conveys the corresponding method signatures.

The implementation is a bit naive as it will grab any and all public
and private methods. Something like Github: seattlerb/sexp_processor.