Generalizing enrichment builder
Instead of the specific use cases, prefer instead the generalized
enrichment builder.

Note, this could be much more modular but for now the enrichment form
building is happening in a singular location; therefore the repeated
knowledge of enrichment types is being isolated to:

* routes
* build_enrichment_form

In a perfect world we'd tease apart the idea of the EnrichmentType,
but that is not yet in place.