Removing incidental database hits
The behavior was such that the current user was being created. Instead
I'm shifting towards using the generated CommandRepositoryInterface.

This means that I'm not checking if methods are called; I could add
expectations, but since the Interface is built from the existing
command module methods, I figure this is adequate.