Adding means for creating todo list items
Given that there is a state related to todo items, I need to have a
means of querying that state to present the todo list to a user. The
most obvious method is to track what to do items are associated with
the work at the given state.

I'm working my way around the todo list/enrichment design, attempting
to resolve how this information will be tracked. Its a multi-
dimensional problem: WorkType, State, Items maps to Work, User, State,
Items.