Adding column names to spec
I keep having to look at the migrations, so I figured I may as well
push this information into the spec so I can see that. Also, I'm not
certain that the column_names will remain as is, so I want a trigger
that will let me know of changes.