Replacing enrichment items with persisted items
Instead of a hard-coded state, I'm now pulling that information from
the database. I'll need to explore how that behaves going forward.

As of today there is quite a bit of REVIEW, TODO, etc that need to be
considered. There is a deprecation warning that is also being printed.
Tomorrow I'll need to review this and attempt to factor things into a
better state.