Encapsulating the finding action object
There are a few options regarding finding the action object. I want to
encapsulate that logic as the current assumption is that all named and
actioned-upon elements in a schema.org/Action are direct children of
the schema.org/Action item declaration.

This is sufficient for now, but will likely not work as well going
forward.