Improving selector behavior for named objects
Instead of making the assumption that the name of the object is an
immediate ancestor of the named object container element, I'm now able
to retrieve the object via a more robust method.

This is likely slower to test, but adds durability to a feature test
that consumes the UI. So while the cost may be high, the rendered HTML
can change without as great of an impact.

See sipity@ebb5d6a for further comments