Further enriching schema.org usage
As we begin to look at acting on forms, it became clear that we needed
to look deeper into what we meant by http://schema.org/Action. The
action is the thing you do. The entryPoint is the context in which you
do it.

Therefore, we teased apart the two related objects. An EntryPoint and
its potentialAction. In the case of the `#take_action_on` method I'm
thinking how can I click on either a submit button or an a-tag.

Since names are helpful, but machines like URLs better, I opted to
first find the URL, and fallback to the named action for something to
click on.