Inverting the potentialAction and entryPoint relation
Given that an action can have entryPoints and an entryPoint can have
possibleActions, which one is the parent container is somewhat
arbitrary.

However, given that a Thing > CreativeWork can have a potentialAction
it makes sense that the Action would contain the entryPoint. In this
way I can have the nested concept, and push the actions into the
CreativeWork container (as they are actions to take on that work).