Moving where clause aspects into join conditions
Its been a long time since I've written SQL outer joins. The key thing
that I had forgotten is that the outer join ON imperative means that
those matching elements are optional. Whereas the WHERE clause must
exist.

I need to write a proper set of tests for this query, as it is
fundamental to much of what I'm doing.