Adding WorkTypes to the database
Given that I may need to drive additional information based on the
WorkType, I want a place that I can begin to describe that
information.

Yes, I could continue to use the magic name, and in some cases I will,
but if I want an association to a ProcessingStrategy, I believe I
should perform my due diligence and craft a proper relationship.