Adjusting dependency injection configuration
To add some clarification. The NetIdValidator#initialize method is
called as part of the Collaborator.validates class method call. It is
configuring the validation rules.

When we call the instance method Collaborator#valid? then the
NetIdValidator#validates_each instance method is called with the
instance of the collaborator, the attribute name, and the value of
that attribute.