Adding ability to conditionally run accessibility
In an effort to ensure accessibility, I'm looking to automate that
behavior using the capybara-accessible gem. Out of the box, these
tests are slower (as they are using Poltergeist to interact via a
headless browser instead of with an HTML page).

I needed to make changes to allow for hidden elements to be found.
I also wanted an option to run choose whether to run with or without
accessibility.