Replacing > with / for Schema.org named actions
As I was working with capybara-accessible, I encountered the problem
where the HTML document was being sent off to Google's accessibility
validation service and would come back altered.

The service returned an HTML document that converted the META content
"todo>required" to "todo&gt;required". As I want to be able to run
these accessibility tests, this is a required change.

https://github.com/Casecommons/capybara-accessible