Replacing event_name a more descriptive name
I'm opting for processing_action_name as it is a reflection of the
processing sub-systme. I would like for it to be :action_name, but
that is a reserved name in the controller. I could introduce a
translation at the controller level, but keeping it the same would
create less confusion.