Capturing who is taking the action
This is a stop-gap in that I am only exposing the requesting actor as
an option instead of both the requesting and on behalf. This is a
future change that we will need to kick around.

However, due to the nature of the database constraints, I want to push
this up as soon as possible so that I can obliterate the staging
database in its infancy instead of requiring a rather confusing
migration (which would be possible).