Querying collaborators required for review
Given that we need to know who requires signoff, this is a
foundational query. By itself it may not have a lot of meaning, but as
I move towards the form regarding advisor signoff, I will need to know
who are all the collaborators for which I will require signoff.