Adding AdvisorSignsOff service
There are some conditionals in this behavior. I want to make sure that
I'm isolating those decisions as the advisor signs off form is already
overworked.

One concern, however, is that I'm driving a querying decision off of
a thing that happens in an unrelated action; The
ProcessingActionForm's save method is registering that the action was
taken. This is buried in the implementation details, and has already
happened.