Adding scenarios to explain state change
By teasing apart the methods, and I hope naming them better, I am
attempting to give a clear indication to the intent of when the state
can advance for an advisor's signoff.

This service object may get folded back into the form, but I'm
concerned that the forms are already doing too much.