Replacing broken query with fixed query
This addresses the issue in which a user has completed one of the
prerequisite actions for a guarded action, but not all of them.

Prior to the fix, if the user completed one prerequisite action then
a guarded action would be available. After this fix a guarded action
is only available if all of the actions have been completed.