Tidying up processing states
When a revision is requested by the grad school, it remains under
review by the grad school; It would be possible to create a separate
state that allows for the Grad Student to make changes and push them
back (or at least go through the pagentry of submiting the item).

This is a future enhancement. To be clear, I want to make sure that
the student is NOT able to edit the item when the Grad School sends it
back for review. Because it would mean the whole list of reviewers
would once again need to review things.