Adding proper filter of processing state for Work
Prior to this change, the queries were filtering on the now non-
existent :processing_state column of the Sipity::Model::Work. This
change makes sure to filter by Processing State name.

Note: I'm opting for "name" instead of ID, because different work
types can have a state with the same name, and from the users
perspective they are the same thing.