Adding config/application.yml to repo
I want a way to share what are the names of the keys our application
is expecting. This also helps in the bootstrap process for our
application.

As a matter of deploys we update the secrets of our application, so
the keys you find within the config/application.yml are not the real
keys.