Content choreography for the work show page
The idea is to suppress areas of the interface when there are no
available decisions or enrichment for a user to make. As previously
noted there are starting to be a lot of decisions in this view. It would
be nice to break them out somehow.

The view structure may be able to be simplified if the state alert
message would get displayed in the same area.