Reworking data model for AccessRights
Given that we are not yet coping with Access Rights, I'm reworking the
data model. I'm not concerned with when it starts; Only concerned with
transitions.

Also, treating the embargo then open access as a first class citizen
in the database. This will ease future development as I look to
refactor the underlying form for access controls.