Reworking LdapClient to ensure coverage
The client was confusing. There were two methods that queried.

One was a method that caught a time out exception. It would then raise
a UserNotFound error. That is confusing, because the exception is now
translating the timeout error to something with a different meaning.

So I've opted to remove that; as it wasn't called. Also, I'm factoring
towards accepting an connection instead of instantiating a new one;
Perhaps there should be a connection that is initialized as part of
boot process of the application. Don't know and not yet important.