Restoring collaborator errors on base
This is an ugly solution, but one to revisit. The validations of the
form throw an error on base if NO collaborators are given that require
review. That error message does belong with any of the collaborators.

It would be possible to push that error onto each of the items. But
we'd need to revisit how we'd do that.