Enforcing authorization for reading attachments
Prior to this commit, the URL for any attachment was publicly
accessible. And while it may have been hard to determine what the URL
should have been, it is something that could easily be disclosed (on
accident).

With this change, authorization is enforced on an attachment. It is a
quick fix and may not reflect the ideal situtation, but it is a better
approximation of what we need.

NOTE: Due to Dragonfly's caching if the policy enforcement changes it
is possible a user that once had access (but should no longer) will
still have access.