Ensuring we only attempt to persist user input
There were a few moving parts in this; First we want to make sure that
we are always rendering enough input fields for the user. Second, due
to changes in how we render collaborators on the form, there was an
unintended side-effect of attempting to persist those collaborators
via the repository.

The solution is to explicitly persist the collaborators_from_input
to the repository; Hence the name changes.

Also clarifying the intent of a guard method:

```ruby
reject_because_an_empty_row_was_submitted_via_user_input?
```

The above method I believe is more clear in its intent.

Closes #311