Renaming mark_as_representative
I have found the name confusing. Mark as representative implies an
action not a property; That is it implies transformative changes. This
is made confusing when we rendering on the view:

```erb
<%= f.input :mark_as_representative %>
```

I would rather clarify by saying:

```erb
<%= f.input :representative_attachment_id %>
```

The latter implies a property, and thus data that is manipulated via a
form object.