Vanity cleanup
Defering to repository methods instead of data structure knowledge.
For me, the jury is still out on this; Is it better to have a data
structure with lots of methods or is it better to leverage a
repository method to convey this information?

Miss you @jimweirich because I believe you would have a helpful
answer.

Closes #353