Addressing comments
Changing from Build Requirements to Goals. This reflects the nature of
the contribution; There is an enforcement that happens in some cases,
but in others its our commitment to know the state of our application.

Code is complex, and each of these metrics are loose approximations of
code health. Taken together they present a picture about the state of
the code. Is this all that we could evaluate? No. There is more. But
this is where we start; Or perhaps more correctly what my expectations
were but are now captured and articulated.

[skip ci]