Enforcing <100% code coverage breaks the build
The 100% is a configurable options.

What I mean by code is `.rb` files in the `app` or `lib` directory.

As this is a goal under discussion (see #372), and one that can be
enforced, I'd like to advocate for automated enforcement.

First: It is a configurable value; If we find that 100% is to
aggressive we can dial that back. We already have 100% coverage, so
keeping to that is less expensive than on code bases we are picking up
to maintain.

Second: My goal for the group is to still push towards Test Drive
Development; I understand the reality of the learning curve for
testing. However, as a reviewer of the code I would like to leverage
any and all tools available to help in the review process. One of
those is making sure that new code is tested; Code coverage is, in my
estimation, a proxy for that.

Third: Similar to Rubocop's enforcement of style guides, this task
again reduces the chatter of pull requests; As a reviewer I will not
need to bug you about tests being written. I can instead focus on
reviewing the code for clarity, consistency, and efficacy.

I am passionate about this but want to make sure I am honoring each of
the team members contributions and perspectives.

So lets discuss this; Is it something you are willing to try for two
sprints? I'm also hear to help you write those tests to reach our
goal.