Guarding closing issue is for clean working dir
This is one means of ensuring that the working directory is clean
(i.e. there are no files in the project that have not been committed).

As part of closing an issue, the plan plan is to run the tests. This
ensures what is tested does not include uncommitted changes.