Conditionally exclusing attachment validation
Given that a person may opt to jump straight to the "Set Access
Controls" without attaching one or more files, we cannot assume that
there is a possible representative attachment.

Given that the HTML5 required syntax gets applied to the select box, I
also need to hide that option if there aren't any attachments to mark
as representative.