Excluding collaborators that have already done it
```
As the grad school reviewer that is signing off on behalf of other
people,
I want to only see a list of people that have not already sign off,
So that I don't end up signing off multiple times for the same person.
```

Prior to this commit, all collaborators were listed. This was a bit
confusing in that I would keep seeing collaborators that had either:

* Had already signed off **or**
* I had already signed off on their behalf