Destroying registered actions on state transition
I'm opting to collaborate with the repository as the query to find
actions to delete is non-trivial. Again, so much indirection, but the
idea is to keep each of the objects tight in purpose.

**Why am I destroying instead of marking as stale?**

The action registry is something transient; That is to say it is used
for determining what actions are available/completed but is not needed
intended for provenance. There is an EventLog that tracks what
happened as should be considered "immutable".