Switching to mysql2 for testing and development
I have encountered two issues with the differences between mysql and
sqlite. I want to make sure that we are vetting those differences in
development so they do not show up in staging.

One issue was the index length of a text field. Sqlite doesn't care
and Mysql needs a length specified. The other issue is since forgetten
by me.