Switching tests to run against SQLite
This is a compromise; There are tests that are failing because the
seeds used for generation are based on assumptions about
auto-incrementing id fields.

I can resolve this by switching the test ecosystem to use mysql AND
adjust the database cleaning to use truncation instead of transaction
(see below for reference). However that bumps the test suite times
from 12 seconds up to 2.5 minutes which is unacceptable.

Again, this is a compromise.

```ruby
RSpec.configure do |config|
  config.before(:suite) do
    DatabaseCleaner.clean_with(:truncation)
  end

  config.before(:each) do
    DatabaseCleaner.strategy = :transaction
  end
end
```