Permitting create_submission_window to work area
In talking with the team, the main purpose of the work area is to
grant permission to create one or more submission windows. What
happens when a submission window is created is dependent on the Work
Area that it is associated with.

The idea is as follows:

When you choose to create a submission window via the
"create_submission_action", the form building object will leverage the
WorkArea to find the appropriate form object for that action.