Refactoring RequestChange forms
This commit got mangled with several ideas; However it contains code
that moves the idea towards a more normalized behavior.

There are a few concerns, though the primary concern is pushing
the request changes towards a common behavior; They are identical in
many ways (aside from on behalf of question).