Working on PowerConverter for WorkArea
The WorkArea is going to inform the module space used for building
forms. As such, I want a way to "quickly" convert an object to the
Work Area.

The main goal is to convert a ProcessingEntity or Work into the
WorkArea, but for now I'm not certain what direction that would be.
There is a clear query path based on the current ERD, but I believe
the ERD may represent the final design as we tease apart the
Submission Window behavior.