Removing external assumption about work submitter
In the case of creating a submission window, I don't want to pass
along to the customized data generator any ideas of who might be the
work submitters.

In doing this, I ended up bootstrapping data such that no one could go
to the `/areas/etd/start` URL because the `:work_submitter` input was
an empty array (`[]`).