Working through work_area routes
As I think about leveraging the WorkArea controller, I want to expose
the common queries and commands within the same controller.

In doing this, I believe I will better understand the interface for
controllers interacting with the processing subsystem.

I'm choosing to separate out the command_action_name vs. the
query_action_name so that the method interfaces are a little easier to
cope with.