Testing the WorkArea command/query actions
As I was working through the refactoring of the processin actions, I
was thinking how to bring the shared behavior together.

It makes sense that the various actions associated with a WorkArea be
connected to the WorkArea controller. Since I have the inflection
point on the runners, I'm able to leverage routes as a way to fine
tune the user facing API and let the controllers be somewhat stupid.

In theory I could get rid of several of the actions and instead focus
on using the runners. But that is a potential future exercise.