Updating access_changes_over_time I18n
Instead of a non-html version, I want to use the translated HTML.
Because we had a key of :access_changes_over_time the SimpleForm
rendering was treating the value as not HTML safe and was escaping
the <i> and </i> tags. By using :access_changes_over_time_html
SimpleForm treats the I18n value as HTML Safe and thus does not escape
the <i> and </i> tags.