Adding WorkType generator to Ulra
This is the preliminary generator, it does not necessarily reflect the
final state of actions. It also conceals a nuanced issue/bug/quirk:

* There are four actions required for completion in order to submit a
  ULRA for review
* Three of those actions can only be taken by the creating user
* One of those actions can only be taken by the assigned advisor

Thus the submit for review button will be available to the person who
did all of the required todo items last. I believe this is acceptable
but perhaps requires a rethinking of the state machine.