Adding WorkArea specific form for create a work
I want to begin associating works with submission windows. This is the
first step saying that the existing form building should go through
the assumed ETD system.

There will need to be a builder going forward that makes a
determination on all forms to find. The existing setup is somewhat
ad-hoc but can be made more sane.