Adding WorkSubmission
Instead of opting for a Work belongs to a SubmissionWindow, I'm
instead preferring that all works are associated with a
WorkSubmission. In doing so, the idea is that a Work can exist outside
of a Submission Window; That is to say if we delete the Submission
Window that the Work was submitted into, we don't want to delete the
Work.