Extracting a proper SubmissionWindowQueries mod
Instead of overworking the WorkAreaQueries, I want a separate query.
This is in preparation for SubmissionWindow action rendering.

Once the SubmissionWindow action rendering is complete, I can bring
that to the ETD work space.

The goal is to move quickly on it so I can add the requisite new
actions.