Apologizing for committing too many things
This commit got away from me; I'm certain that I've introduced issues.
The reason it got away from me is that I was exploring the interplay
of view rendering; And that is fraught with peril in Rails (at least
how I wanted to do it).

So apologies to future maintainers for this change. I will be
performing additional refactoring before I submit a pull request.