Adding a base class to the given form
In adding a base class to the given form, I'm able to delegate the
translations down to the associated model. However, due to inheritance
concerns and crazy error messaging, I needed to break inheritance.

The reason for breaking inheritance, was that the base_class and
delegation for ActiveModel::Validations means all kinds of bad things
happen upon inheritance.