Ensuring that signoffs are dual registered
Given that people can do things on behalf of other people, when there
are two options presented, we need to make sure that both options are
registered.

The scenario that this solves is:

Given that Collaborator 1 and Collaborator 2 are required to approve
of a work (directly or via a delegate) for the state to advance.

1. Collaborator 1 approves
2. Someone approves on behalf of Collaborator 2

The state should advance but did not. However if you reverse the steps
it would've advanced the state.