Adding script to find uncalled methods
This is not a failsafe tool, but leverages the always insightful Ryan
Davis's work via the debrid gem.

The idea being that to accept ownership, we need to not only ensure
proper code coverage but make sure that the methods in production
are called outside of tests.

Ruby makes this complicate. Rails makes this complicated. Method
missing makes this complicated, but the methods identified by debride
are candidates for review.