Adjusting dashboard permissions query
Based on testing, when more than one person had created a work, that
gave them the ability to see (in the Dashboard) all of the records.
This is both undesired and confusing.

The underlying permissions were still enforce (i.e. "Show"-ing the
object would fail).

The resulting solution was to normalize how both the strategy and
entity level permissions are built. Its now an even crazier than
normal query building. But it works.