Adding primary key to sipity_work_submissions
Prior to this commit the following Ruby code threw an exception:

```ruby
> Sipity::Models::WorkSubmission.where(work: work).destroy_all
=> !! #<ActiveRecord::StatementInvalid: Mysql2::Error: Unknown column
   'sipity_work_submissions.' in 'where clause': DELETE FROM
   `sipity_work_submissions` WHERE `sipity_work_submissions`.`` =
   NULL>
```

This is because I didn't have a primary key associated with the table.
I could've opted to use the more appropriate composite_primary_keys
gem, however that adds lots of additional code for a very small gain.
Yes, it would be better to have this be a composite key, but I don't
need to add hundreds of lines of code for this edge case.

https://github.com/composite-primary-keys/composite_primary_keys